+++ examples/support-vector-machines.c
@@ -5,14 +5,14 @@
 int
 main(int argc, const char *argv[])
 {
-    const double x_data[] =
+    const float x_data[] =
     {
         1.0, 2.0, 3.0, 4.0,
         4.0, 3.0, 2.0, 1.0
     };
     LeMatrix *x = le_matrix_new_from_data(2, 4, x_data);
     
-    const double y_data[] =
+    const float y_data[] =
     {
         0.0, 0.0, 1.0, 1.0
     };
